move postion state related files to global directory
