+++ src/modules/map/components/zoomButtons/ZoomToExtentButton.js
@@ -2,7 +2,7 @@ import { html } from 'lit-html';
 import { BaElement } from '../../../BaElement';
 import css from './zoomToExtentButton.css';
 import { $injector } from '../../../../injection';
-import { setFit } from '../../store/position.action';
+import { setFit } from '../../../../store/position/position.action';
 
 /**
  * Button that zooms map to extent
