+++ src/modules/map/injection/index.js
@@ -1,5 +1,5 @@
 import { LayersPlugin } from '../store/LayersPlugin';
-import { PositionPlugin } from '../store/PositionPlugin';
+import { PositionPlugin } from '../../../store/position/PositionPlugin';
 import { OlMeasurementHandler } from '../components/olMap/handler/measure/OlMeasurementHandler';
 import { OlGeolocationHandler } from '../components/olMap/handler/geolocation/OlGeolocationHandler';
 import { GeolocationPlugin } from '../store/GeolocationPlugin';
