+++ src/store/position/PositionPlugin.js
@@ -1,6 +1,6 @@
-import { $injector } from '../../../injection';
-import { QueryParameters } from '../../../services/domain/queryParameters';
-import { BaPlugin } from '../../../store/BaPlugin';
+import { $injector } from '../../injection';
+import { QueryParameters } from '../../services/domain/queryParameters';
+import { BaPlugin } from '../BaPlugin';
 import { changeZoomAndCenter, setFit } from './position.action';
 
 /**
