+++ test/modules/map/components/zoomButtons/ZoomToExtentButton.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 
-import { positionReducer } from '../../../../../src/modules/map/store/position.reducer';
+import { positionReducer } from '../../../../../src/store/position/position.reducer';
 import { TestUtils } from '../../../../test-utils.js';
 import { $injector } from '../../../../../src/injection';
 import { fromLonLat } from 'ol/proj';
