+++ test/service/ShareService.test.js
@@ -1,8 +1,8 @@
 import { $injector } from '../../src/injection';
 import { addLayer } from '../../src/modules/map/store/layers.action';
 import { layersReducer } from '../../src/modules/map/store/layers.reducer';
-import { changeZoomAndCenter } from '../../src/modules/map/store/position.action';
-import { positionReducer } from '../../src/modules/map/store/position.reducer';
+import { changeZoomAndCenter } from '../../src/store/position/position.action';
+import { positionReducer } from '../../src/store/position/position.reducer';
 import { setCurrent } from '../../src/modules/topics/store/topics.action';
 import { topicsReducer } from '../../src/modules/topics/store/topics.reducer';
 import { QueryParameters } from '../../src/services/domain/queryParameters';
